refactor: result -> game 패키지로 변경 및 Game 클래스 위치를 game으로 변경
