+++ src/main/java/blackjack/domain/game/Result.java
@@ -1,4 +1,4 @@
-package blackjack.domain.result;
+package blackjack.domain.game;
 
 public enum Result {
     WIN("승"),
