+++ src/test/java/blackjack/domain/game/GameTest.java
@@ -1,9 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardSymbol;
 import blackjack.domain.card.Deck;
+import blackjack.domain.game.Game;
 import blackjack.domain.gameplayer.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
