+++ src/test/java/blackjack/domain/game/ResultRefereeTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.result;
+package blackjack.domain.game;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
